Fix: started services folder and deleted postSchema from post model
