+++ client/store/coordinates.js
@@ -19,7 +19,7 @@ export const getCoordinatesThunk = () => {
         // 'http://localhost:8080/api/photos/photo',
         //'http://192.168.1.152:8080/api/photos/photo',
         //anna's ip address: 192.168.1.152
-        `http://192.168.1.6:8080/api/maps/coordinates`
+        `http://110.0.0.153:8080/api/maps/coordinates`
       );
       dispatch(getCoordinates(data));
     } catch (error) {
