+++ server/db/models/post.js
@@ -32,42 +32,4 @@ const Post = db.define('post', {
   },
 });
 
-const PostSchema = {
-  title: {
-    type: Sequelize.STRING,
-    allowNull: false,
-    validate: {
-      notEmpty: true,
-    },
-  },
-  description: {
-    type: Sequelize.TEXT,
-    allowNull: false,
-    validate: {
-      notEmpty: true,
-    },
-  },
-  latitude: {
-    type: Sequelize.FLOAT,
-    allowNull: false,
-    validate: {
-      notEmpty: true,
-    },
-  },
-  longitude: {
-    type: Sequelize.FLOAT,
-    allowNull: false,
-    validate: {
-      notEmpty: true,
-    },
-  },
-  primaryUrl: {
-    type: Sequelize.TEXT,
-    allowNull: false,
-    validate: {
-      notEmpty: true,
-    },
-  },
-};
-
 module.exports = Post;
